+++ data/mongod.lock
@@ -1,1 +1,1 @@
-15223
+4253
