--- src/models/config.ts
@@ -1,3 +1,0 @@
-export interface Config {
-	port?: number;
-}
