+++ src/models/users.js
@@ -1,0 +1,22 @@
+const mongoose = require("mongoose");
+
+const userSchema = new mongoose.Schema({
+	username: {
+		type: String,
+		required: true,
+		unique: true,
+	},
+	hash: {
+		type: String,
+		required: true,
+	},
+	salt: {
+		type: String,
+		required: false,
+	},
+});
+const User = mongoose.model("user", userSchema);
+
+module.exports = {
+	User
+};
