+++ src/routes/index.js
@@ -1,0 +1,7 @@
+const express = require("express");
+
+const router = express.Router();
+
+router.get("/", (req, res) => res.render("index"));
+
+module.exports = router;
