--- src/routes/index.ts
@@ -1,7 +1,0 @@
-import express from "express";
-
-const router = express.Router();
-
-router.get("/", (req, res) => res.render("index"));
-
-export default router;
