+++ src/utils/common.js
@@ -1,0 +1,22 @@
+const cryptojs = require("crypto").webcrypto;
+
+async function sha256(message) {
+	// encode as UTF-8
+	const msgBuffer = new TextEncoder().encode(message);
+
+	// hash the message
+	const hashBuffer = await cryptojs.subtle.digest("SHA-256", msgBuffer);
+
+	// convert ArrayBuffer to Array
+	const hashArray = Array.from(new Uint8Array(hashBuffer));
+
+	// convert bytes to hex string
+	const hashHex = hashArray
+		.map((b) => ("00" + b.toString(16)).slice(-2))
+		.join("");
+	return hashHex;
+}
+
+module.exports = {
+	sha256
+};
